+++ src/main/java/com/iesam/digLibrary/features/loans/domain/EndLoanUseCase.java
@@ -9,7 +9,7 @@ public class EndLoanUseCase {
             this.repository = repository;
         }
 
-        public void execute(int loanId){
+        public void execute(long loanId){
             Loan loan = repository.getLoanById(loanId);
             Date currentDate = new Date();
             Loan loanFinished = new Loan(loan.loanId, loan.resource, loan.user, loan.loanDate,loan.expectedDate ,currentDate);
